+++ libs/mini-rx-store/src/lib/models.ts
@@ -33,9 +33,15 @@ export interface StoreConfig<T> {
     extensions: StoreExtension[];
 }
 
+// Used for the Redux API: Store.feature
 export interface FeatureStoreConfig<StateType> {
     initialState: StateType;
     metaReducers?: MetaReducer<StateType>[];
+}
+
+// Used for createFeatureStore, new FeatureStore
+export interface FeatureStoreInstanceConfig {
+    multi?: boolean;
 }
 
 export class Actions extends Observable<Action> {}